Java: готово

+++ jApp/src/main/java/j/s/yarlykov/data/provider/ForecastProvider.java
@@ -15,16 +15,15 @@ import j.s.yarlykov.R;
 import j.s.yarlykov.data.domain.CityForecast;
 import j.s.yarlykov.data.domain.Forecast;
 import j.s.yarlykov.data.network.WeatherDataLoader;
-import j.s.yarlykov.util.Utils;
 
-import static j.s.yarlykov.data.domain.CityForecast.*;
+import static j.s.yarlykov.data.domain.CityForecast.MeteoData;
 
 public class ForecastProvider {
 
     private static ForecastProvider instance;
 
     public static ForecastProvider getInstance() {
-        if(instance == null) {
+        if (instance == null) {
             instance = new ForecastProvider();
         }
         return instance;
@@ -33,7 +32,7 @@ public class ForecastProvider {
     public CityForecast getRealForecast(Context context, String city) {
         final JSONObject jsonObject = WeatherDataLoader.getJSONData(city);
 
-        if(jsonObject != null) {
+        if (jsonObject != null) {
             try {
                 JSONObject details = jsonObject.getJSONArray("weather").getJSONObject(0);
                 JSONObject main = jsonObject.getJSONObject("main");
@@ -42,8 +41,8 @@ public class ForecastProvider {
                 return new CityForecast(
                         extractPlaceName(jsonObject),
                         extractIconResourceId(context, details),
-                        (int)extractCurrentTemp(main),
-                        (int)extractWindSpeed(wind),
+                        extractCurrentTemp(main),
+                        extractWindSpeed(wind),
                         extractHumidity(main),
                         extractPressure(main)
                 );
@@ -58,23 +57,23 @@ public class ForecastProvider {
         return jsonObject.getString("name");
     }
 
-    private double extractCurrentTemp(JSONObject main) throws JSONException {
-        return main.getDouble("temp");
+    private int extractCurrentTemp(JSONObject main) throws JSONException {
+        return (int) main.getDouble("temp");
     }
 
-    private int extractHumidity(JSONObject main)throws JSONException {
+    private int extractHumidity(JSONObject main) throws JSONException {
         return main.getInt("humidity");
     }
 
-    private int extractPressure(JSONObject main)throws JSONException {
+    private int extractPressure(JSONObject main) throws JSONException {
         return main.getInt("pressure");
     }
 
-    private double extractWindSpeed(JSONObject wind)throws JSONException {
-        return wind.getDouble("speed");
+    private float extractWindSpeed(JSONObject wind) throws JSONException {
+        return (float) wind.getDouble("speed");
     }
 
-    private String extractWindDirect(JSONObject wind)throws JSONException {
+    private String extractWindDirect(JSONObject wind) throws JSONException {
         int deg = wind.getInt("deg");
         return "Stub";
     }
@@ -84,7 +83,6 @@ public class ForecastProvider {
         Resources resources = context.getResources();
         return resources.getIdentifier(icon, "drawable",
                 context.getPackageName());
-
     }
 
     /**
@@ -95,9 +93,9 @@ public class ForecastProvider {
 
     private ForecastProvider() {
         forecasts = Arrays.asList(
-                new Forecast(R.drawable.rain,12, 10, 89, 750),
-                new Forecast(R.drawable.sunny,25, 5, 73, 771),
-                new Forecast(R.drawable.sun,18, 4, 85, 769),
+                new Forecast(R.drawable.rain, 12, 10, 89, 750),
+                new Forecast(R.drawable.sunny, 25, 5, 73, 771),
+                new Forecast(R.drawable.sun, 18, 4, 85, 769),
                 new Forecast(R.drawable.snow, -3, 1, 59, 741),
                 new Forecast(R.drawable.cloud2, 7, 2, 51, 749),
                 new Forecast(R.drawable.cloud1, 11, 6, 64, 742)
@@ -115,7 +113,7 @@ public class ForecastProvider {
     /**
      * TODO: Получить прогноз с интересующими данными
      */
-    public CityForecast getForecastCustom(String city, Set<MeteoData> request){
+    public CityForecast getForecastCustom(String city, Set<MeteoData> request) {
         return new CityForecast(city, forecasts.get(index())).clearUnused(request);
     }
 
@@ -132,7 +130,7 @@ public class ForecastProvider {
     /**
      * TODO: Сгенерить рандомный индекс массива
      */
-    private int index(){
+    private int index() {
         Random randomForecast = new Random();
         return randomForecast.nextInt(forecasts.size());
     }
